Use a tester div to check if CSS added
When injectStyles is called, a div is added as a child to the plotly
div. Next, check for whether the (entirely arbitrary) styles defined in
the plotcss have been applied to the div, if they have, assume the CSS
has been injected, otherwise, inject the CSS. After that, delete the
test div.